tests: move sfinae to its own header file
